Generic function to create the index.md file to accompany the data deposit package

Generic function to create the index.md file to accompany the data deposit package

Title

Usage

make_index(
  scheme,
  path = ".",
  overwrite = FALSE,
  template = scheme_path_index_template(),
  author = NULL,
  make = c("html", "pdf"),
  pandoc_bin = "pandoc",
  pandoc_args = "-s",
  ...
)

# S3 method for dmdSchemeSet
make_index(
  scheme,
  path = ".",
  overwrite = FALSE,
  template = scheme_path_index_template(),
  author = NULL,
  make = c("pdf", "html"),
  pandoc_bin = "pandoc",
  pandoc_args = "-s",
  ...
)

Arguments

scheme

a dmdScheme from which the values for the tokens in the template should be taken

path

path to where the `index` should be created. The file name of the created index is identical to the file name of the template.

overwrite

if TRUE, the target index file will be overwritten automatically, unless the target index is equal to the template, in which case, an error would be raised in all cases.

template

template to be used for the index file. For details see the vignette Create and Customize the index template. The default template is at system.file("index.md", package = "dmdScheme")

author

of the index file

make

character vector containing types into which the generated index file should be converted to. default is html and pdf. This function uses pandoc for the conversion!.

pandoc_bin

pandoc executable. Needs fully qualified path when not in $PATH.

pandoc_args

arguments for calling pandoc

...

additional arguments for methods

Value

returns path to the created index.md file

Examples

if (FALSE) {
  # takes to long for CRAN
  make_index( dmdScheme_example(), path = tempdir() )
}