Create examples in working directory

Each package based on a dmdScheme can contain examples. This function is the interface to these examples. In the package dmdScheme, no examples are included. The function has two basic usages:

by using make_example(schemeName = "NameOfTheScheme") all included examples are listed

by using make_example(name = "basic", schemeName = "NameOfTheScheme") it will create the example named basic in a subdirectory in the current working directory. An existing directory with the same name, will nod be overwritten!

Usage

make_example(name)

Arguments

name

name of the example

Value

invisibly NULL

Details

The examples have to be located in a directory called example_data. The function is doing two things:

Copying the complete directory from the example_data directory to the current working directory

running knitr::purl on all ./code/*.Rmd to extract the code into .R script files. If you want to include an RMarkdown files in the ./code directory from thisa, use the .rmd extension (small letters).

Examples

make_example()
#> Included examples are:
#> 
if (FALSE) {
make_example("basic")
}