Open the metadata scheme as a spreadsheet in a spreadsheet editor

Open system.file(paste0(schemeName, ".xlsx"), package = schemeName) in excel. New data can be entered and the file has to be saved at a different location as it is a read-only file.

open_new_spreadsheet(
  file = NULL,
  open = TRUE,
  keepData = FALSE,
  format = TRUE,
  overwrite = FALSE,
  verbose = FALSE
)

Arguments

file
if not NULL, the template will be saved to this file.

open
if TRUE, the file will be opened. This can produce different results depending on the OS, browsr and browser settings.

keepData
if TRUE the data entry areas will be emptied. If FALSE. the example data will be included.

format
if FALSE the sheet will be opened as the sheet is. if TRUE, it will be formated nicely.

overwrite
if TRUE, the file specified in file will be overwritten. if FALSE, an error will be raised ehen the file exists.

verbose
give verbose progress info. Useful for debugging.

Value

invisibly the fully qualified path to the file which would have been opened, if open == TRUE.

Examples

if (FALSE) {
open_new_spreadsheet(schemeName = "dmdScheme", format = FALSE, verbose = TRUE)
}