Generic function for creating a report from an object x

This generic function creates a report based on the object.

report(
  x,
  file = tempfile(),
  open = TRUE,
  report = "html",
  report_author = "Myself",
  report_title = "Report of something",
  ...
)

# S3 method for character
report(
  x,
  file = tempfile(),
  open = TRUE,
  report = "html",
  report_author = "Tester",
  report_title = "Validation of data against dmdScheme",
  ...
)

# S3 method for dmdScheme_validation
report(
  x,
  file = tempfile(),
  open = TRUE,
  report = "html",
  report_author = "Tester",
  report_title = "Validation of data against dmdScheme",
  ...
)

Arguments

x
object of which the report should be creted used to select a method

file
name of the file containing the generated report, including extension. If missing, ot will be saved as a temporary file in the temporary folder.

open
if TRUE, open the report. Default: TRUE

report
determines if and in which format a report of the validation should be generated. Allowed values are:

none: no report is generated

html: a html (.html) report is generated and opened

pdf : a pdf (.pdf) report is generated and opened

word: a word (.docx)report is generated and opened Additional values can be implemented by the different methods and will be documented in the Details section.

report_author
name of the author to be included in the report

report_title
title of the report to be included in the report

...
further arguments passed to or from other methods

Value

return the path and filename of the report

Details

report.character creates a report of the object returnes from a validate().

report.dmdScheme_validation creates a report of the object returnes from a validate().

Methods (by class)

character: report of a dmdScheme_validation object.

dmdScheme_validation: report of a dmdScheme_validation object.

Examples

## Report of `dmdScheme_validation`
report(scheme_path_xlsx())

#> Validating Experiment

#> Validating Genus

#> Validating Treatments

#> Validating Measurement

#> Validating DataExtraction

#> Validating DataFileMetaData

#> 
#> 
#> processing file: validation_report.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.....                                                                 |   7%
#>    inline R code fragments
#> 
#>   |                                                                              |.........                                                             |  13%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |..............                                                        |  20%
#>    inline R code fragments
#> 
#>   |                                                                              |...................                                                   |  27%
#> label: unnamed-chunk-1
#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |.................................                                     |  47%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................                                 |  53%
#> label: unnamed-chunk-3 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: unnamed-chunk-4 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |...................................................                   |  73%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: unnamed-chunk-5 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |.............................................................         |  87%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................................................................     |  93%
#> label: info
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: validation_report.knit.md

#> /usr/local/bin/pandoc +RTS -K512m -RTS validation_report.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output validation_report.html --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /usr/local/lib/R/4.0/site-library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --number-sections --variable 'theme:bootstrap' --include-in-header /var/folders/50/wcr5bjwn75q595n6x82gxj280000gq/T//RtmppAXifF/rmarkdown-str505c2a8fca31.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --lua-filter /usr/local/lib/R/4.0/site-library/rmarkdown/rmd/lua/pagebreak.lua --lua-filter /usr/local/lib/R/4.0/site-library/rmarkdown/rmd/lua/latex-div.lua 

#> 
#> Output created: validation_report.html

#> Non-interactive session. If interactive session, the report would be opened.

#> [1] "/var/folders/50/wcr5bjwn75q595n6x82gxj280000gq/T//RtmppAXifF/file505c46236ccb"

## Report of `dmdScheme_validation`
report(validate(dmdScheme_raw()))

#> Validating Experiment

#> Validating Genus

#> Validating Treatments

#> Validating Measurement

#> Validating DataExtraction

#> Validating DataFileMetaData

#> 
#> 
#> processing file: validation_report.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.....                                                                 |   7%
#>    inline R code fragments
#> 
#>   |                                                                              |.........                                                             |  13%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |..............                                                        |  20%
#>    inline R code fragments
#> 
#>   |                                                                              |...................                                                   |  27%
#> label: unnamed-chunk-1
#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |.................................                                     |  47%
#>   ordinary text without R code
#> 
#>   |                                                                              |.....................................                                 |  53%
#> label: unnamed-chunk-3 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: unnamed-chunk-4 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |...................................................                   |  73%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: unnamed-chunk-5 (with options) 
#> List of 1
#>  $ results: chr "asis"
#> 
#>   |                                                                              |.............................................................         |  87%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................................................................     |  93%
#> label: info
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: validation_report.knit.md

#> /usr/local/bin/pandoc +RTS -K512m -RTS validation_report.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash+smart --output validation_report.html --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /usr/local/lib/R/4.0/site-library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --number-sections --variable 'theme:bootstrap' --include-in-header /var/folders/50/wcr5bjwn75q595n6x82gxj280000gq/T//RtmppAXifF/rmarkdown-str505c76f5b3f2.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --lua-filter /usr/local/lib/R/4.0/site-library/rmarkdown/rmd/lua/pagebreak.lua --lua-filter /usr/local/lib/R/4.0/site-library/rmarkdown/rmd/lua/latex-div.lua 

#> 
#> Output created: validation_report.html

#> Non-interactive session. If interactive session, the report would be opened.

#> [1] "/var/folders/50/wcr5bjwn75q595n6x82gxj280000gq/T//RtmppAXifF/file505c7cf911e1"

if (FALSE) {
report(
  x = dmdScheme_raw(),
  report = "html",
  report_author = "The Author I am",
  report_title = "A Nice Report"
)
}