Generic function to convert the data stored in the object x into a new object of class dmdScheme_raw...

Generic function to convert the data stored in the object x into a new object of class dmdScheme_raw...

Usage

as_dmdScheme_raw(x, ...)

# S3 method for dmdSchemeData
as_dmdScheme_raw(x, ...)

# S3 method for dmdSchemeSet
as_dmdScheme_raw(x, ...)

# S3 method for xml_document
as_dmdScheme_raw(x, useSchemeInXml = NULL, ...)

Arguments

x

object to be converted

...

additional arguments for methods

useSchemeInXml

if TRUE, use scheme definition in xml and raise an error if the xml does not contain a scheme definition. If False, use the scheme definition from the corresponding installed package, even if the xml contains a scheme definition. if NULL (the default), use the definition in the xml if it contains a definition, if not use the corresponding definition from the installed package.

Value

dmdScheme as object of class dmdScheme_raw

Examples

as_dmdScheme_raw(dmdScheme(), keepData = TRUE)
#> $MdBibliometric
#>       propertySet    valueProperty      type
#> 1  MdBibliometric       uploadType character
#> 2            <NA>              doi character
#> 3            <NA>  publicationDate character
#> 4            <NA>            title character
#> 5            <NA>      description character
#> 6            <NA>          version character
#> 7            <NA>         language character
#> 8            <NA>         keywords character
#> 9            <NA>  additionalNotes character
#> 10           <NA>     accessRights character
#> 11           <NA> accessRightsInfo character
#> 12           <NA>          license character
#>                                   suggestedValues
#> 1                                            <NA>
#> 2                                            <NA>
#> 3                                            <NA>
#> 4                                            <NA>
#> 5                                            <NA>
#> 6                                            <NA>
#> 7                                            <NA>
#> 8                                            <NA>
#> 9                                            <NA>
#> 10                                           <NA>
#> 11                                           <NA>
#> 12 Creative Commons Attribution 4.0 International
#>                            allowedValues
#> 1  Dataset, Image - Photo, Video / Audio
#> 2                                   <NA>
#> 3                                   <NA>
#> 4                                   <NA>
#> 5                                   <NA>
#> 6                                   <NA>
#> 7                                   <NA>
#> 8                                   <NA>
#> 9                                   <NA>
#> 10   Open, Embargoed, Restricted, Closed
#> 11                                  <NA>
#> 12                                  <NA>
#>                                                                                                                 Description
#> 1                                                    The type of the data. Needs tobe one as specified in the allowedValues
#> 2                                                                         The DOI of the dataset. Usually the reserved DOI.
#> 3                                                          Date of publication of the data package in the format YYYY-MM-DD
#> 4                                                                                                 Title of the data package
#> 5                                                                                   Short description of the data deposited
#> 6  The version of the dataset. Usually empty, but the version should be specified if the dataset is augmented or corrected.
#> 7                                                                          The language of the dataset. E.g. "eng", "fr", …
#> 8                                                                                           Keywords describing the dataset
#> 9                                                                 Some additional info, which does not fit in any category.
#> 10                                                                   Access rights to the data, normally Open or Embargoed.
#> 11 if Embargoed, the date until embargoed (YYYY-MM-DD), if Restricted, conditions under which access is granted to the data
#> 12                                               Should normally be "Creative Commons Attribution 4.0 International" (CCA4)
#>    DATA
#> 1  <NA>
#> 2  <NA>
#> 3  <NA>
#> 4  <NA>
#> 5  <NA>
#> 6  <NA>
#> 7  <NA>
#> 8  <NA>
#> 9  <NA>
#> 10 <NA>
#> 11 <NA>
#> 12 <NA>
#> 
#> $MdAuthors
#>       propertySet                    MdAuthors                      ...3
#> 1   valueProperty                     authorID                familyName
#> 2            unit                         <NA>                      <NA>
#> 3            type                    character                 character
#> 4 suggestedValues                         <NA>                      <NA>
#> 5   allowedValues                         <NA>                      <NA>
#> 6     Description Just an id - can be a number Family name of the author
#> 7            DATA                         <NA>                      <NA>
#> 8   MULTIPLE ROWS                         <NA>                      <NA>
#>                           ...4                      ...5
#> 1                    givenName               affiliation
#> 2                         <NA>                      <NA>
#> 3                    character                 character
#> 4                         <NA>                      <NA>
#> 5                         <NA>                      <NA>
#> 6 Given name(s) of the authors Affiliation of the author
#> 7                         <NA>                      <NA>
#> 8                         <NA>                      <NA>
#>                                                                         ...6
#> 1                                                                      orcid
#> 2                                                                       <NA>
#> 3                                                                  character
#> 4                                                                       <NA>
#> 5                                                                       <NA>
#> 6 ORCID (Open Researcher and Contributor ID, as issued by https://orcid.org)
#> 7                                                                       <NA>
#> 8                                                                       <NA>
#>                                                                                                                                                                                                                                            ...7
#> 1                                                                                                                                                                                                                                          role
#> 2                                                                                                                                                                                                                                          <NA>
#> 3                                                                                                                                                                                                                                     character
#> 4                                                                                                                                                                                                                                          <NA>
#> 5 Conceptualization, Data curation, Formal Analysis, Funding acquisition, Investigation, Methodology, Project administration, Resources, Software, Supervision, Validation, Visualization, Writing – original draft, Writing – review & editing
#> 6                                                                                                                        Role of the author, using the CRediT – Contributor Roles Taxonomy at https://casrai.org/credit/. Can be more than one.
#> 7                                                                                                                                                                                                                                          <NA>
#> 8                                                                                                                                                                                                                                          <NA>
#> 
#> $Experiment
#>    propertySet      valueProperty unit      type
#> 1   Experiment               name <NA> character
#> 2         <NA>        temperature <NA> character
#> 3         <NA>              light <NA> character
#> 4         <NA>           humidity <NA> character
#> 5         <NA>          incubator <NA> character
#> 6         <NA>          container <NA> character
#> 7         <NA>    microcosmVolume   ml   numeric
#> 8         <NA>          mediaType <NA> character
#> 9         <NA> mediaConcentration  g/l   numeric
#> 10        <NA>  cultureConditions <NA> character
#> 11        <NA>       comunityType <NA> character
#> 12        <NA>     mediaAdditions <NA> character
#> 13        <NA>           duration days   integer
#> 14        <NA>            comment <NA> character
#>                                            suggestedValues
#> 1                                                     <NA>
#> 2               treatment, in degrees celsius, measurement
#> 3              treatment,light, dark, cycle , e.g. 16:8 LD
#> 4                        treatment, relative humidity in %
#> 5                                              none, bench
#> 6                                                     <NA>
#> 7                                                     <NA>
#> 8                                                     <NA>
#> 9                                                     <NA>
#> 10                                    axenic, dirty, clean
#> 11 treatment, single trophic level, multiple trophic level
#> 12                                                    <NA>
#> 13                                                    <NA>
#> 14                                                    <NA>
#>                                                                                                                  Description
#> 1                                                                                                The name of the experiment.
#> 2  Temperature used for all treatments. If different between treatments, use "treatment" and specify in the Treatment sheet.
#> 3        Light used for all treatments. If different between treatments, use "treatment" and specify in the Treatment sheet.
#> 4     Humidity used for all treatments. If different between treatments, use "treatment" and specify in the Treatment sheet.
#> 5                                                                                            What type of incubator is used.
#> 6                                                                                            What type of container is used.
#> 7                                                   Volume of the microcosm container. Not the volume of the culture medium!
#> 8                                                                                                                       <NA>
#> 9                                                                                                                       <NA>
#> 10                                                                            Conditions of the cultures for all treatments.
#> 11                                                                                Characterisation of the microbe community.
#> 12                                                                                                                      <NA>
#> 13                 Length of the experiment in days. This should only include the time in which the measurements were taken!
#> 14                                                                 Additional features of the Experiment you want to provide
#>    DATA dmdScheme v0.9.9
#> 1                   <NA>
#> 2                   <NA>
#> 3                   <NA>
#> 4                   <NA>
#> 5                   <NA>
#> 6                   <NA>
#> 7                   <NA>
#> 8                   <NA>
#> 9                   <NA>
#> 10                  <NA>
#> 11                  <NA>
#> 12                  <NA>
#> 13                  <NA>
#> 14                  <NA>
#> 
#> $Genus
#>       propertySet
#> 1   valueProperty
#> 2            unit
#> 3            type
#> 4 suggestedValues
#> 5   allowedValues
#> 6     Description
#> 7            DATA
#> 8   MULTIPLE ROWS
#>                                                            Genus
#> 1                                                      speciesID
#> 2                                                           <NA>
#> 3                                                      character
#> 4                                                           <NA>
#> 5                                                           <NA>
#> 6 Id of the species and strain. Each speciesId has to be unique.
#> 7                                                           <NA>
#> 8                                                           <NA>
#>                                   ...3
#> 1                               colour
#> 2                                 <NA>
#> 3                            character
#> 4                                 <NA>
#> 5                                 <NA>
#> 6 Where the species was obtained from.
#> 7                                 <NA>
#> 8                                 <NA>
#>                                                                                                                            ...4
#> 1                                                                                                                       density
#> 2                                                                                                                    cells / ml
#> 3                                                                                                                     character
#> 4                                                                                                                     treatment
#> 5                                                                                                                          <NA>
#> 6 Initial density used for all treatments. If different between treatments, use "treatment" and specify in the Treatment sheet.
#> 7                                                                                                                          <NA>
#> 8                                                                                                                          <NA>
#>                                          ...5      ...6
#> 1                             functionalGroup   comment
#> 2                                        <NA>      <NA>
#> 3                                   character character
#> 4 bacteria, bacterivore, predator, phototroph      <NA>
#> 5                                        <NA>      <NA>
#> 6            Functional group of the species.      <NA>
#> 7                                        <NA>      <NA>
#> 8                                        <NA>      <NA>
#> 
#> $Treatments
#>       propertySet
#> 1   valueProperty
#> 2            unit
#> 3            type
#> 4 suggestedValues
#> 5   allowedValues
#> 6     Description
#> 7            DATA
#> 8   MULTIPLE ROWS
#>                                                                                                                               Treatments
#> 1                                                                                                                            treatmentID
#> 2                                                                                                                                   <NA>
#> 3                                                                                                                              character
#> 4         species, temperatur, light, initial density, comunity composition, densities, dispersal, viscosity, disturbance, communityType
#> 5                                                                                                                                   <NA>
#> 6 ID of the the treatment decribed in this a row. Each treatmentId can occur multiple times as it can contain multiple treatment levels.
#> 7                                                                                                                                   <NA>
#> 8                                                                                                                                   <NA>
#>                                                                                                                                                                                  ...3
#> 1                                                                                                                                                                treatmentLevelHeight
#> 2                                                                                                                                                                                <NA>
#> 3                                                                                                                                                                           character
#> 4                                                                                                                                                          value,  variable: freetext
#> 5                                                                                                                                                                                <NA>
#> 6 The value of the parameter if the parameter is constant over time, or a description of the variability. If unit is speciesId, comma separated list of all species in the treatment.
#> 7                                                                                                                                                                                <NA>
#> 8                                                                                                                                                                                <NA>
#>        ...4
#> 1   comment
#> 2      <NA>
#> 3 character
#> 4      <NA>
#> 5      <NA>
#> 6      <NA>
#> 7      <NA>
#> 8      <NA>
#> 
#> $Measurement
#>       propertySet
#> 1   valueProperty
#> 2            unit
#> 3            type
#> 4 suggestedValues
#> 5   allowedValues
#> 6     Description
#> 7            DATA
#> 8   MULTIPLE ROWS
#>                                                                                                                                                                                                                  Measurement
#> 1                                                                                                                                                                                                              measurementID
#> 2                                                                                                                                                                                                                       <NA>
#> 3                                                                                                                                                                                                                  character
#> 4                                                                                                                                                                                                                       <NA>
#> 5                                                                                                                                                                                                                       <NA>
#> 6 Id of the Measurement process. This includes methodology, variables . Each measurementId specifies one Measurement process and must be unique in this column. Should be in the mapping column in the DataFileMetaData tab.
#> 7                                                                                                                                                                                                                       <NA>
#> 8                                                                                                                                                                                                                       <NA>
#>                                                    ...3
#> 1                                              variable
#> 2                                                  <NA>
#> 3                                             character
#> 4 O2 concentration, video, manual count, abundance, DNA
#> 5                                                  <NA>
#> 6                                The variable measured.
#> 7                                                  <NA>
#> 8                                                  <NA>
#>                         ...4                          ...5
#> 1                     method                          unit
#> 2                       <NA>                          <NA>
#> 3                  character                     character
#> 4 presens Optode, microscopy                %, mmol, count
#> 5                       <NA>                          <NA>
#> 6   Name of the method used. Unit of the measured variable
#> 7                       <NA>                          <NA>
#> 8                       <NA>                          <NA>
#>                                                                                                                        ...6
#> 1                                                                                                                    object
#> 2                                                                                                                      <NA>
#> 3                                                                                                                 character
#> 4                                                                                  species, OUT, gene, community, particles
#> 5                                                                                                                      <NA>
#> 6 The object measured. E.g. species in the case of manual count, gene for genetic analysis, particle for particle counters.
#> 7                                                                                                                      <NA>
#> 8                                                                                                                      <NA>
#>                                              ...7
#> 1                         noOfSamplesInTimeSeries
#> 2                                            <NA>
#> 3                                         integer
#> 4                                            <NA>
#> 5                                            <NA>
#> 6 Total number of all samples in the time series.
#> 7                                            <NA>
#> 8                                            <NA>
#>                                                                                                                                                      ...8
#> 1                                                                                                                                          samplingVolume
#> 2                                                                                                                                                      ml
#> 3                                                                                                                                                 numeric
#> 4                                                                                                                                                    <NA>
#> 5                                                                                                                                                    <NA>
#> 6 The sampling volumne. If e.g. atmosphere in container is sampled (oxygen measurements), than enter 0.\r\nPlease use NA if sampling volumne is variable.
#> 7                                                                                                                                                    <NA>
#> 8                                                                                                                                                    <NA>
#>                                                           ...9
#> 1                                             dataExtractionID
#> 2                                                         <NA>
#> 3                                                    character
#> 4                                                         <NA>
#> 5                                                         <NA>
#> 6 as used in the sheet DataExtraction, column dataExtractionID
#> 7                                                         <NA>
#> 8                                                         <NA>
#>                                                                                                            ...10
#> 1                                                                                                   measuredFrom
#> 2                                                                                                           <NA>
#> 3                                                                                                      character
#> 4                                                                                                           <NA>
#> 5                                                                                                           <NA>
#> 6 if measured from the experiment, raw, else the measurementId (first column) of the Measurement it is based on.
#> 7                                                                                                           <NA>
#> 8                                                                                                           <NA>
#>       ...11
#> 1   comment
#> 2      <NA>
#> 3 character
#> 4      <NA>
#> 5      <NA>
#> 6      <NA>
#> 7      <NA>
#> 8      <NA>
#> 
#> $DataExtraction
#>       propertySet
#> 1   valueProperty
#> 2            unit
#> 3            type
#> 4 suggestedValues
#> 5   allowedValues
#> 6     Description
#> 7            DATA
#> 8   MULTIPLE ROWS
#>                                                                                                                                                                                           DataExtraction
#> 1                                                                                                                                                                                       dataExtractionID
#> 2                                                                                                                                                                                                   <NA>
#> 3                                                                                                                                                                                              character
#> 4                                                                                                                                                                                                   <NA>
#> 5                                                                                                                                                                                                   <NA>
#> 6 Name of the DataExtraction process. This includes methodology, variables . Each name specifies one extraction process and can occur multiple times in the case of multiple parameters in the analysis.
#> 7                                                                                                                                                                                                   <NA>
#> 8                                                                                                                                                                                                   <NA>
#>                                                                                                     ...3
#> 1                                                                                                 method
#> 2                                                                                                   <NA>
#> 3                                                                                              character
#> 4                                                                                           bemovi x.y.z
#> 5                                                                                                   <NA>
#> 6 Method used for the DataExtraction process. If possible including version (in the case of R packages).
#> 7                                                                                                   <NA>
#> 8                                                                                                   <NA>
#>                                                                                   ...4
#> 1                                                                            parameter
#> 2                                                                                 <NA>
#> 3                                                                            character
#> 4                                                                                 <NA>
#> 5                                                                                 <NA>
#> 6 parameter in the analysis. Only needs to be specified if it varies from the default.
#> 7                                                                                 <NA>
#> 8                                                                                 <NA>
#>                                                        ...5      ...6
#> 1                                                     value   comment
#> 2                                                      <NA>      <NA>
#> 3                                                 character character
#> 4                                                      <NA>      <NA>
#> 5                                                      <NA>      <NA>
#> 6 value of the parameter (you can enter a number or a word)      <NA>
#> 7                                                      <NA>      <NA>
#> 8                                                      <NA>      <NA>
#> 
#> $DataFileMetaData
#>       propertySet          DataFileMetaData
#> 1   valueProperty              dataFileName
#> 2            unit                      <NA>
#> 3            type                 character
#> 4 suggestedValues                      <NA>
#> 5   allowedValues                      <NA>
#> 6     Description the name of the data set.
#> 7            DATA                      <NA>
#> 8   MULTIPLE ROWS                      <NA>
#>                                                                                                                                                                           ...3
#> 1                                                                                                                                                                   columnName
#> 2                                                                                                                                                                         <NA>
#> 3                                                                                                                                                                    character
#> 4                                                                                                                                                                         <NA>
#> 5                                                                                                                                                                         <NA>
#> 6 Name of column in the data file. \r\nEach column in the data file needs to be documented!\r\nor NA\r\nif it is for the whole data file and not specified in the dataFileName
#> 7                                                                                                                                                                         <NA>
#> 8                                                                                                                                                                         <NA>
#>                                                                                                                                                                                                                              ...4
#> 1                                                                                                                                                                                                                      columnData
#> 2                                                                                                                                                                                                                            <NA>
#> 3                                                                                                                                                                                                                       character
#> 4                                                                                                                                                                                                                            <NA>
#> 5                                                                                                                                                                                      ID, Treatment, Measurement, Species, other
#> 6 The type of the data in the column.\r\nID: ID field (unique ID of unit of replication);\r\nTreatment: specifies treatment; \r\nMeasurement: contains measurements; \r\nSpecies: contains species; \r\nother: other type of data
#> 7                                                                                                                                                                                                                            <NA>
#> 8                                                                                                                                                                                                                            <NA>
#>                                                                                                                                                                                                                                                                ...5
#> 1                                                                                                                                                                                                                                                     mappingColumn
#> 2                                                                                                                                                                                                                                                              <NA>
#> 3                                                                                                                                                                                                                                                         character
#> 4                                                                                                                                                                                                                                                              <NA>
#> 5                                                                                                                                                                                                                                                              <NA>
#> 6 columnData = Treatment: \r\ntreatmentID as in the Treatment tab; \r\ncolumnData = Species: \r\ntreatmentID refering to species composition as in the Treatment tab\r\ncolumnData = Measurement: \r\nmeasurementID as in the Measurement tab; \r\notherwise:\r\nNA
#> 7                                                                                                                                                                                                                                                              <NA>
#> 8                                                                                                                                                                                                                                                              <NA>
#>                                                         ...6
#> 1                                                       type
#> 2                                                       <NA>
#> 3                                                  character
#> 4                                                       <NA>
#> 5 integer, numeric, character, logical, datetime, date, time
#> 6                                        Type of the column.
#> 7                                                       <NA>
#> 8                                                       <NA>
#>                                                                                                                                                                                                                                                                                                                                ...7
#> 1                                                                                                                                                                                                                                                                                                                       description
#> 2                                                                                                                                                                                                                                                                                                                              <NA>
#> 3                                                                                                                                                                                                                                                                                                                         character
#> 4                                                                                                                                                                                                                                                                                                                              <NA>
#> 5                                                                                                                                                                                                                                                                                                                              <NA>
#> 6 if column contains measurement: General description. If type is datatime, date, or time, give the order of year month day hour minute second as e.g. ymdhms, ymd, or hms. (Do not give any other information, e.g. give nothing about how months are entered (e.g. number or name), or how years, months, day, etc are separated.
#> 7                                                                                                                                                                                                                                                                                                                              <NA>
#> 8                                                                                                                                                                                                                                                                                                                              <NA>
#>        ...8
#> 1   comment
#> 2      <NA>
#> 3 character
#> 4      <NA>
#> 5      <NA>
#> 6      <NA>
#> 7      <NA>
#> 8      <NA>
#> 
#> attr(,"propertyName")
#> [1] "dmdScheme"
#> attr(,"dmdSchemeName")
#> [1] "dmdScheme"
#> attr(,"dmdSchemeVersion")
#> [1] "0.9.9"
#> attr(,"fileName")
#> [1] "dmdScheme.xlsx"
#> attr(,"class")
#> [1] "dmdSchemeSet_raw" "list"