Object of class dmdSchemeSet containing the authorative definition of the dmdScheme.

The dataset contains the authorative definition of the dmdScheme. It contains no data, except one row NA. There are two S3 classes defined and used:

dmdSchemeData:

a data.frame with the class dmdSchemeData. Each column is one property of the metadata. It has the following attributes:

propertyName:

name of the data. In the spreadsheet, it is in the cell below (in the Experiment tab) or to the right (other tabs).

unit:

the unit of the data in each column.

type:

the type of the data in the column. These will be validated in the validate function. See there for details.

suggestedValues:

suggested values for the data of each column. These will be validated in the validate function. See there for details.

allowedValues:

allowed values for the data of each column. These will be validated in the validate function. See there for details.

Description:

general description of the columns.

names:

the names of the columns.

dmdSchemeSet:

a list with where each element is a dmdSchemeData object with additional attributes:

propertyName:

name of the dmdScheme used. In the spreadsheet, it is in the cell H1 in ther in the Experiment tab (DATA dmeScheme v0.9.9.)

dmdSchemeVersion:

version of the dmdScheme used. In the spreadsheet, it is in the cell H1 in ther in the Experiment tab (DATA dmeScheme v0.9.9.)

names:

the names of the dmdSchemeData sets. In the spreadsheet, the names of the tabs.

Usage

dmdScheme()

Examples

dmdScheme()
#>  dmdScheme - dmdSchemeSet
#>      MdBibliometric - dmdSchemeData
#> A       Names : uploadType | doi | publicationDate | title | description | version | language | keywords | additionalNotes | accessRights | accessRightsInfo | license
#> A       Units : 
#> A       Type  : character | character | character | character | character | character | character | character | character | character | character | character
#> 
#>           uploadType  doi publicationDate title description version language
#> D       1       <NA> <NA>            <NA>  <NA>        <NA>    <NA>     <NA>
#>           keywords additionalNotes accessRights accessRightsInfo license
#> D       1     <NA>            <NA>         <NA>             <NA>    <NA>
#> 
#>      MdAuthors - dmdSchemeData
#> A       Names : authorID | familyName | givenName | affiliation | orcid | role
#> A       Units : NA | NA | NA | NA | NA | NA
#> A       Type  : character | character | character | character | character | character
#> 
#>           authorID familyName givenName affiliation orcid role
#> D       1     <NA>       <NA>      <NA>        <NA>  <NA> <NA>
#> 
#>      Experiment - dmdSchemeData
#> A       Names : name | temperature | light | humidity | incubator | container | microcosmVolume | mediaType | mediaConcentration | cultureConditions | comunityType | mediaAdditions | duration | comment
#> A       Units : NA | NA | NA | NA | NA | NA | ml | NA | g/l | NA | NA | NA | days | NA
#> A       Type  : character | character | character | character | character | character | numeric | character | numeric | character | character | character | integer | character
#> 
#>           name temperature light humidity incubator container microcosmVolume
#> D       1 <NA>        <NA>  <NA>     <NA>      <NA>      <NA>              NA
#>           mediaType mediaConcentration cultureConditions comunityType
#> D       1      <NA>                 NA              <NA>         <NA>
#>           mediaAdditions duration comment
#> D       1           <NA>       NA    <NA>
#> 
#>      Genus - dmdSchemeData
#> A       Names : speciesID | colour | density | functionalGroup | comment
#> A       Units : NA | NA | cells / ml | NA | NA
#> A       Type  : character | character | character | character | character
#> 
#>           speciesID colour density functionalGroup comment
#> D       1      <NA>   <NA>    <NA>            <NA>    <NA>
#> 
#>      Treatments - dmdSchemeData
#> A       Names : treatmentID | treatmentLevelHeight | comment
#> A       Units : NA | NA | NA
#> A       Type  : character | character | character
#> 
#>           treatmentID treatmentLevelHeight comment
#> D       1        <NA>                 <NA>    <NA>
#> 
#>      Measurement - dmdSchemeData
#> A       Names : measurementID | variable | method | unit | object | noOfSamplesInTimeSeries | samplingVolume | dataExtractionID | measuredFrom | comment
#> A       Units : NA | NA | NA | NA | NA | NA | ml | NA | NA | NA
#> A       Type  : character | character | character | character | character | integer | numeric | character | character | character
#> 
#>           measurementID variable method unit object noOfSamplesInTimeSeries
#> D       1          <NA>     <NA>   <NA> <NA>   <NA>                      NA
#>           samplingVolume dataExtractionID measuredFrom comment
#> D       1             NA             <NA>         <NA>    <NA>
#> 
#>      DataExtraction - dmdSchemeData
#> A       Names : dataExtractionID | method | parameter | value | comment
#> A       Units : NA | NA | NA | NA | NA
#> A       Type  : character | character | character | character | character
#> 
#>           dataExtractionID method parameter value comment
#> D       1             <NA>   <NA>      <NA>  <NA>    <NA>
#> 
#>      DataFileMetaData - dmdSchemeData
#> A       Names : dataFileName | columnName | columnData | mappingColumn | type | description | comment
#> A       Units : NA | NA | NA | NA | NA | NA | NA
#> A       Type  : character | character | character | character | character | character | character
#> 
#>           dataFileName columnName columnData mappingColumn type description
#> D       1         <NA>       <NA>       <NA>          <NA> <NA>        <NA>
#>           comment
#> D       1    <NA>
#>